requirements.txt
fastapi
uvicorn
pydantic
torch
httpx
groq
python-dotenv